+++ src/main/java/com/main/easyFix/appuser/AppUserRepository.java
@@ -11,8 +11,8 @@ import java.util.Optional;
 
 @Repository
 @Transactional(readOnly = true)
-public interface AppUserRepository extends JpaRepository<AppUser, Long> {
-  Optional<AppUser> findById(Long id);
+public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
+  Optional<AppUser> findById(int id);
   Optional<AppUser> findByEmail(String email);
   List<AppUser> findAll();
